Split V9 fact compilation by evidence domain
Decompose safety-score-v9-fact-set.ts (4296 lines, 493-line max function, 332 branches) into eight evidence-domain modules - schema, context, operational resilience, backing, exit, control, peg/supply, and wrapper - with the root module preserving all ten original exports and the trusted-extension guard. Pure code motion verified by sorted-body diff; 149 pinned scoring and golden-card tests pass with zero fixture changes. The reserve-exposure key helper lives in the schema module so the sibling graph stays acyclic. Audit findings R-13/Q-06/S-11 (P3-04 tranche 2).
